Créer un objet de classe 'commune' à partir d'un data frame

Cette fonction prend un data frame contenant des informations sur une commune et valide que toutes les lignes correspondent à une seule commune (en vérifiant que le Code de la commune est unique). Elle transforme ensuite ce data frame en un objet de classe commune pour lui attribuer un comportement spécifique lors de son utilisation.

Usage

creer_commune(df)

Arguments

df

Un data frame contenant des informations sur les élus d'une commune. Le data frame doit inclure la colonne Code de la commune.

Value

Un objet de classe commune contenant les mêmes informations que le data frame d'entrée, mais avec une nouvelle classe attribuée.

Examples

# Exemples d'utilisation

# Données (villes)

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")
#> Error: object 'Libellé de la commune' not found

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")
#> Error: object 'Libellé de la commune' not found

# Utilisation de la fonction

df_Nantes <- creer_commune(df_Nantes)
#> Error: object 'df_Nantes' not found

df_Aignan <- creer_commune(df_Aignan)
#> Error: object 'df_Aignan' not found