Trouver l'élu(e) le plus âgé(e) et le/la plus jeune

Cette fonction permet de trouver l'élu(e) le plus âgé(e) et le/la plus jeune dans un data frame d'élus. Elle calcule l'âge des élus à partir de leur date de naissance, puis retourne les informations concernant l'élu(e) le/la plus âgé(e) et le/la plus jeune. La fonction trie les élus par date de naissance et sélectionne le premier et le dernier élu.

Usage

trouver_elu_le_plus_age_jeune(df)

Arguments

df

Un data frame contenant les informations sur les élus, incluant les colonnes Nom de l'élu, Prénom de l'élu, Date de naissance, et Libellé de la commune.

Value

Un tibble avec les informations de l'élu(e) le/la plus âgé(e) et de l'élu(e) le/la plus jeune. Les colonnes retournées incluent :

Nom de l'élu : Le nom de l'élu,

Prénom de l'élu : Le prénom de l'élu,

Date de naissance : La date de naissance de l'élu,

âge : L'âge calculé de l'élu,

Libellé de la commune : La commune à laquelle appartient l'élu.

Examples

# Exemples d'utilisation

# Données (villes ou départements)

df_Nantes <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Nantes")
#> Error: object 'Libellé de la commune' not found

df_Aignan <- df_Gers_Loire_Atlantique |>
  filter(`Libellé de la commune` == "Aignan")
#> Error: object 'Libellé de la commune' not found

df_Loire_Atlantique <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Loire-Atlantique")
#> Error: object 'Libellé du département' not found

df_Gers <- df_Gers_Loire_Atlantique |>
  filter(`Libellé du département` == "Gers")
#> Error: object 'Libellé du département' not found

# Utilisation de la fonction

trouver_elu_le_plus_age_jeune(df_Nantes)
#> Error in trouver_elu_le_plus_age_jeune(df_Nantes): could not find function "trouver_elu_le_plus_age_jeune"

trouver_elu_le_plus_age_jeune(df_Aignan)
#> Error in trouver_elu_le_plus_age_jeune(df_Aignan): could not find function "trouver_elu_le_plus_age_jeune"

trouver_elu_le_plus_age_jeune(df_Loire_Atlantique)
#> Error in trouver_elu_le_plus_age_jeune(df_Loire_Atlantique): could not find function "trouver_elu_le_plus_age_jeune"

trouver_elu_le_plus_age_jeune(df_Gers)
#> Error in trouver_elu_le_plus_age_jeune(df_Gers): could not find function "trouver_elu_le_plus_age_jeune"